Separated the style of the number of cart items out.
For style:
I added a _cart.less file in public/css, so the element with
id="cart-count" will use the style defined there.

For data:
In layout.hbs, <span id="cart-count" data-bind="text:
cartCount"></span>, (I changed the id from "cart" to "cart-count",
forget this unimportant detail), the value of the data, which is the
number of items in the cart is associated with "cartCount" that is defined
in scripts/viewModels/HomeVM.js.

For control:
In all the place where this number should be updated, I get the reference to this variable and update it.