Added order history display
While implementing, to reuse the client side model, I changed the name
of some of the fields of the document stored in the user table when the
order is completed: e.g., changed "order_details" to "items".

Bug: search for "wild", add two of each of the results, click cart icon,
the displayed quantity is 3 and 1, which should be 2 and 2.

Questionable behavior: if a user changes quantity or removes
items, and then searches something and comes back, the changed order isn't saved, consider save orders while a user does modification.